+++ aerospike/__init__.py
@@ -1,4 +1,3 @@
 # aerospike extension module will be stored in this folder (i.e package)
 # Same with the exception and predicates extension modules
 from .aerospike import *
-from . import exception, predicates
