+++ setup.py
@@ -282,7 +282,7 @@ setup(
     ext_modules=[
         Extension(
             # Extension Name
-            'aerospike.aerospike',
+            'aerospike',
 
             # Source Files
             [
@@ -369,38 +369,6 @@ setup(
                 'src/main/client/metrics.c'
             ],
 
-            # Compile
-            include_dirs=include_dirs,
-            extra_compile_args=extra_compile_args,
-
-            # Link
-            library_dirs=library_dirs,
-            libraries=libraries,
-            extra_objects=extra_objects,
-            extra_link_args=extra_link_args,
-        ),
-        Extension(
-            name="aerospike.exception",
-            sources=[
-                'src/main/exception.c',
-            ],
-            # Compile
-            include_dirs=include_dirs,
-            extra_compile_args=extra_compile_args,
-
-            # Link
-            library_dirs=library_dirs,
-            libraries=libraries,
-            extra_objects=extra_objects,
-            extra_link_args=extra_link_args,
-        ),
-        Extension(
-            name="aerospike.predicates",
-            sources=[
-                'src/main/predicates.c',
-                'src/main/exception.c',
-                'src/main/geospatial/dumps.c'
-            ],
             # Compile
             include_dirs=include_dirs,
             extra_compile_args=extra_compile_args,
