+++ src/include/predicates.h
@@ -18,7 +18,4 @@
 
 #include <Python.h>
 
-PyObject *AerospikeException_New(void);
-void raise_exception(as_error *err);
-PyObject *raise_exception_old(as_error *err);
-void remove_exception(as_error *err);
+PyObject *AerospikePredicates_New(void);
