+++ src/main/conversions.c
@@ -54,6 +54,12 @@
 #define PY_KEYT_KEY 2
 #define PY_KEYT_DIGEST 3
 
+#define PY_EXCEPTION_CODE 0
+#define PY_EXCEPTION_MSG 1
+#define PY_EXCEPTION_FILE 2
+#define PY_EXCEPTION_LINE 3
+#define AS_PY_EXCEPTION_IN_DOUBT 4
+
 #define CTX_KEY "ctx"
 #define CDT_CTX_ORDER_KEY "order_key"
 #define CDT_CTX_PAD_KEY "pad_key"
@@ -2381,6 +2387,40 @@ as_status metadata_to_pyobject(as_error *err, const as_record *rec,
 
     *obj = py_meta;
     return err->code;
+}
+
+void error_to_pyobject(const as_error *err, PyObject **obj)
+{
+    PyObject *py_file = NULL;
+    if (err->file) {
+        py_file = PyUnicode_FromString(err->file);
+    }
+    else {
+        Py_INCREF(Py_None);
+        py_file = Py_None;
+    }
+    PyObject *py_line = NULL;
+    if (err->line > 0) {
+        py_line = PyLong_FromLong(err->line);
+    }
+    else {
+        Py_INCREF(Py_None);
+        py_line = Py_None;
+    }
+
+    PyObject *py_code = PyLong_FromLongLong(err->code);
+    PyObject *py_message = PyUnicode_FromString(err->message);
+
+    PyObject *py_in_doubt = err->in_doubt ? Py_True : Py_False;
+    Py_INCREF(py_in_doubt);
+
+    PyObject *py_err = PyTuple_New(5);
+    PyTuple_SetItem(py_err, PY_EXCEPTION_CODE, py_code);
+    PyTuple_SetItem(py_err, PY_EXCEPTION_MSG, py_message);
+    PyTuple_SetItem(py_err, PY_EXCEPTION_FILE, py_file);
+    PyTuple_SetItem(py_err, PY_EXCEPTION_LINE, py_line);
+    PyTuple_SetItem(py_err, AS_PY_EXCEPTION_IN_DOUBT, py_in_doubt);
+    *obj = py_err;
 }
 
 void initialize_bin_for_strictypes(AerospikeClient *self, as_error *err,
