+++ src/main/exception.c
@@ -19,12 +19,16 @@
 #include <aerospike/as_error.h>
 #include <aerospike/as_status.h>
 
+#include "conversions.h"
 #include <string.h>
 #include <stdlib.h>
 #include "exceptions.h"
 #include "exception_types.h"
+#include "macros.h"
 
-PyObject *PyInit_exception(void)
+static PyObject *module;
+
+PyObject *AerospikeException_New(void)
 {
     static struct PyModuleDef moduledef = {PyModuleDef_HEAD_INIT,
                                            "aerospike.exception",
@@ -34,7 +38,7 @@ PyObject *PyInit_exception(void)
                                            NULL,
                                            NULL,
                                            NULL};
-    PyObject *module = PyModule_Create(&moduledef);
+    module = PyModule_Create(&moduledef);
 
     struct exceptions exceptions_array;
 
@@ -550,22 +554,22 @@ PyObject *PyInit_exception(void)
     return module;
 }
 
+void remove_exception(as_error *err)
+{
+    PyObject *py_key = NULL, *py_value = NULL;
+    Py_ssize_t pos = 0;
+    PyObject *py_module_dict = PyModule_GetDict(module);
+
+    while (PyDict_Next(py_module_dict, &pos, &py_key, &py_value)) {
+        Py_DECREF(py_value);
+    }
+}
+
 void raise_exception(as_error *err)
 {
     PyObject *py_key = NULL, *py_value = NULL;
     Py_ssize_t pos = 0;
-
-    // PyObject *py_importlib_module = PyImport_ImportModule("aerospike.exception");
-
-    // PyObject *py_importlib_module = PyImport_ImportModule("importlib");
-    // PyObject *py_exception_module = PyObject_CallMethod(
-    //     py_importlib_module, "import_module", "%s", "aerospike.exception");
-
-    PyObject *py_module_name = PyUnicode_FromString("aerospike.exception");
-    PyObject *py_exception_module = PyImport_GetModule(py_module_name);
-
-    PyObject *py_module_dict = PyModule_GetDict(py_exception_module);
-    Py_DECREF(py_exception_module);
+    PyObject *py_module_dict = PyModule_GetDict(module);
     bool found = false;
 
     while (PyDict_Next(py_module_dict, &pos, &py_key, &py_value)) {
@@ -636,11 +640,7 @@ PyObject *raise_exception_old(as_error *err)
 {
     PyObject *py_key = NULL, *py_value = NULL;
     Py_ssize_t pos = 0;
-    PyObject *py_module_name = PyUnicode_FromString("aerospike.exception");
-    PyObject *py_exception_module = PyImport_GetModule(py_module_name);
-
-    PyObject *py_module_dict = PyModule_GetDict(py_exception_module);
-    Py_DECREF(py_exception_module);
+    PyObject *py_module_dict = PyModule_GetDict(module);
     bool found = false;
 
     while (PyDict_Next(py_module_dict, &pos, &py_key, &py_value)) {
@@ -692,38 +692,4 @@ PyObject *raise_exception_old(as_error *err)
         }
     }
     return py_value;
-}
-
-void error_to_pyobject(const as_error *err, PyObject **obj)
-{
-    PyObject *py_file = NULL;
-    if (err->file) {
-        py_file = PyUnicode_FromString(err->file);
-    }
-    else {
-        Py_INCREF(Py_None);
-        py_file = Py_None;
-    }
-    PyObject *py_line = NULL;
-    if (err->line > 0) {
-        py_line = PyLong_FromLong(err->line);
-    }
-    else {
-        Py_INCREF(Py_None);
-        py_line = Py_None;
-    }
-
-    PyObject *py_code = PyLong_FromLongLong(err->code);
-    PyObject *py_message = PyUnicode_FromString(err->message);
-
-    PyObject *py_in_doubt = err->in_doubt ? Py_True : Py_False;
-    Py_INCREF(py_in_doubt);
-
-    PyObject *py_err = PyTuple_New(5);
-    PyTuple_SetItem(py_err, PY_EXCEPTION_CODE, py_code);
-    PyTuple_SetItem(py_err, PY_EXCEPTION_MSG, py_message);
-    PyTuple_SetItem(py_err, PY_EXCEPTION_FILE, py_file);
-    PyTuple_SetItem(py_err, PY_EXCEPTION_LINE, py_line);
-    PyTuple_SetItem(py_err, AS_PY_EXCEPTION_IN_DOUBT, py_in_doubt);
-    *obj = py_err;
 }
